requirements-dev.txt
-r requirements.txt

rpdb==0.2.0
tqdm==4.67.1
ipython==8.32.0
flake8==7.1.1
pylint==3.3.4
pep8==1.7.1
django-silk==5.3.2
pytest-cov==6.0.0
pytest-django==4.9.0
pytest-factoryboy==2.7.0
Faker==35.2.0
django-cors-headers==4.7.0
httpx==0.28.1
eth-account==0.13.7
web3==7.8.0
bip-utils==2.9.3

